Split executor module
